Refactor vocab module, rename subword vocabs.
FinalfusionSubwordVocab -> BucketSubwordVocab,
FinalfusionNGramVocab -> ExplicitSubwordVocab, NGramIndexer ->
ExplicitIndexer. Move SimpleVocab, SubwordVocab, VocabWrap to
their own modules. Fix up trait bounds for Vocabs.
